Harden toolchain guard and crash marker; split AppModel
- Add scripts/check-toolchain.sh and wire it into verify/test/app/install targets so a Command Line Tools-only developer directory fails with the exact fix instead of cryptic SwiftUIMacros errors.
- Make ManualControlMarker writes crash-durable (same-directory temp file + fsync + rename + directory fsync), return success, and surface a status warning when the manual-mode recovery marker cannot be written.
- Split AppModel.swift (2612 -> 526 lines) into seven same-target extension files by concern (Lifecycle, Polling, Control, Profiles, Fans, MenuBar, Notifications); pure relocation, no behavior change.
- Update MakefileTrustGateTests and ManualControlMarkerTests for the new contracts.

Verified: make verify green (1344 tests, 0 failures), warnings-as-errors build, app bundling, plist/codesign, viftyctl identifier checks.
