+++ Tests/ViftyCoreTests/MakefileTrustGateTests.swift
@@ -5,10 +5,10 @@ final class MakefileTrustGateTests: XCTestCase {
     func testVerifyTargetRunsLocalTrustGates() throws {
         let makefile = try read("Makefile")
 
-        XCTAssertTrue(makefile.contains("verify: ## Run fast local trust gates without installing"))
+        XCTAssertTrue(makefile.contains("verify: check-toolchain ## Run fast local trust gates without installing"))
         XCTAssertTrue(makefile.contains("verify-full: verify release-contract-ruby-tests installer-lifecycle-ruby-tests ui-review-ruby-tests ## Run full trust gates, including slow XCTest suites, for CI/release-facing checks"))
-        XCTAssertTrue(makefile.contains("test-fast: ## Run the fast local XCTest suite"))
-        XCTAssertTrue(makefile.contains("test-full: ## Run the full XCTest suite, including slow evidence/release script tests"))
+        XCTAssertTrue(makefile.contains("test-fast: check-toolchain ## Run the fast local XCTest suite"))
+        XCTAssertTrue(makefile.contains("test-full: check-toolchain ## Run the full XCTest suite, including slow evidence/release script tests"))
         XCTAssertTrue(makefile.contains("repair-helper: ## Explicitly repair the installed privileged helper"))
         XCTAssertTrue(makefile.contains("validation-evidence: ## Collect read-only release/hardware validation evidence"))
         XCTAssertTrue(makefile.contains("validation-evidence-current-build: ## Build current app and collect read-only local-ad-hoc validation evidence"))
@@ -163,15 +163,15 @@ final class MakefileTrustGateTests: XCTestCase {
         let makefile = try read("Makefile")
 
         XCTAssertTrue(makefile.contains("install-dev-adhoc repair-helper uninstall-helper"))
-        XCTAssertTrue(makefile.contains("run-app: ## Build and open the local app bundle"))
+        XCTAssertTrue(makefile.contains("run-app: check-toolchain ## Build and open the local app bundle"))
         XCTAssertTrue(makefile.contains("install-public-release: ## Verify and install the exact current published release archive"))
         XCTAssertTrue(makefile.contains("PUBLIC_RELEASE_ARCHIVE is required and must be an absolute path to the canonical public release zip"))
         XCTAssertTrue(makefile.contains("CONFIGURATION=release ./scripts/install-vifty.sh --public-release-archive \"$(PUBLIC_RELEASE_ARCHIVE)\""))
-        XCTAssertTrue(makefile.contains("verify: ## Run fast local trust gates without installing"))
+        XCTAssertTrue(makefile.contains("verify: check-toolchain ## Run fast local trust gates without installing"))
         XCTAssertTrue(makefile.contains("verify-full: verify release-contract-ruby-tests installer-lifecycle-ruby-tests ui-review-ruby-tests ## Run full trust gates, including slow XCTest suites, for CI/release-facing checks"))
         XCTAssertTrue(makefile.contains("repair-helper: ## Explicitly repair the installed privileged helper"))
         XCTAssertTrue(makefile.contains("./scripts/repair-vifty-helper.sh --app \"$(REPAIR_HELPER_APP)\""))
-        XCTAssertTrue(makefile.contains("install-dev-adhoc: ## Explicit debug-only install with exact UID/path XPC allowlist"))
+        XCTAssertTrue(makefile.contains("install-dev-adhoc: check-toolchain ## Explicit debug-only install with exact UID/path XPC allowlist"))
         XCTAssertTrue(makefile.contains("VIFTY_ENABLE_ADHOC_XPC=1 ./scripts/install-vifty.sh"))
         XCTAssertTrue(makefile.contains("uninstall-helper: ## Safely remove the installed privileged helper"))
         XCTAssertTrue(makefile.contains("validation-evidence: ## Collect read-only release/hardware validation evidence"))
